Factor out generateRequest into a lib/ai/openai module
